Extract shared SageCoffeeEntity base class
Add entity.py with SageCoffeeEntity(CoordinatorEntity) that centralises
the DeviceInfo block, _attr_has_entity_name, and _handle_coordinator_update
which were duplicated verbatim across all six platform files.

All entity classes now inherit from SageCoffeeEntity in addition to their
HA platform base, removing the boilerplate from each __init__ and keeping
a single place to update device metadata in future.

Closes #14
