+++ custom_components/sagecoffee/number.py
@@ -11,13 +11,12 @@ from homeassistant.components.number import (
     NumberEntity,
     NumberEntityDescription,
 )
-from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers.device_registry import DeviceInfo
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from . import SageCoffeeConfigEntry, SageCoffeeCoordinator
-from .const import DOMAIN, STATE_ASLEEP
+from .const import STATE_ASLEEP
+from .entity import SageCoffeeEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -75,11 +74,10 @@ async def async_setup_entry(
     async_add_entities(entities)
 
 
-class SageCoffeeNumber(CoordinatorEntity[SageCoffeeCoordinator], NumberEntity):
+class SageCoffeeNumber(SageCoffeeEntity, NumberEntity):
     """Represents a Sage Coffee number entity."""
 
     entity_description: SageCoffeeNumberEntityDescription
-    _attr_has_entity_name = True
 
     def __init__(
         self,
@@ -88,20 +86,9 @@ class SageCoffeeNumber(CoordinatorEntity[SageCoffeeCoordinator], NumberEntity):
         description: SageCoffeeNumberEntityDescription,
     ) -> None:
         """Initialize the number entity."""
-        super().__init__(coordinator)
+        super().__init__(coordinator, appliance)
         self.entity_description = description
-        self._appliance = appliance
-        self._serial = appliance.serial_number
         self._attr_unique_id = f"{self._serial}_{description.key}"
-
-        # Device info
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, self._serial)},
-            name=appliance.name or f"Sage Coffee {self._serial[-4:]}",
-            manufacturer="Sage/Breville",
-            model=appliance.model or "Unknown",
-            serial_number=self._serial,
-        )
 
     @property
     def native_value(self) -> float | None:
